:tada: First commit ! Got some things to work (gnome/linux dev newbie)
Displays a window with a headerbar and a box layout, with a picture and a label.
By default the window is server-side-decorated, an action in the menu allows
switching to CSD. The image on the left is loaded from base64 encoding in
a text file, like in Triakr eID files - not there yet with XML though.
Struct definitions for eID data.